Added check in reset_running_commands and cleaned some tests.

+++ smartdispatch/command_manager.py
@@ -42,13 +42,14 @@ class CommandManager(object):
                 self._move_line_between_files(running_commands_file, finished_commands_file, command + '\n')
 
     def reset_running_commands(self):
-        with utils.open_with_lock(self._commands_filename, 'r+') as commands_file:
-            with utils.open_with_lock(self._running_commands_filename, 'r+') as running_commands_file:
-                commands = running_commands_file.readlines()
-                if len(commands) > 0:
-                    running_commands_file.seek(0, os.SEEK_SET)
-                    running_commands_file.truncate()
+        if os.isfile(self._running_commands_filename):
+            with utils.open_with_lock(self._commands_filename, 'r+') as commands_file:
+                with utils.open_with_lock(self._running_commands_filename, 'r+') as running_commands_file:
+                    commands = running_commands_file.readlines()
+                    if len(commands) > 0:
+                        running_commands_file.seek(0, os.SEEK_SET)
+                        running_commands_file.truncate()
 
-                    commands += commands_file.readlines()
-                    commands_file.seek(0, os.SEEK_SET)
-                    commands_file.writelines(commands)
+                        commands += commands_file.readlines()
+                        commands_file.seek(0, os.SEEK_SET)
+                        commands_file.writelines(commands)
